test/emacs: Add tests for `message-cite-reply-position'
`message-cite-reply-position' affects the position of the email body
and signature relative to the email citation in email replies.

Test each of the three possible values of
`message-cite-reply-position'. If `message-cite-reply-position' is
'traditional or 'below, place the email signature below the
citation. If `message-cite-reply-position' is 'above, place the email
signature above the citation (as in Gmail-style email replies).