python: move legacy python bindings to contrib.
This forces us to verify the build and test suite completes without
the legacy python bindings.

As of this commit the legacy python bindings are unsupported by
notmuch. They are provided in contrib/ for now, but will most likely
eventually be removed from there as well. We recommend that
downstream/distro packagers phase out their packages for the legacy
python bindings.