Set 'untrusted-content' to t in all modes
Instead of trying to guess in which modes this is safe not to set in,
let's just set it in all of them.

Background (added by db):

untrusted-content was added in Emacs 29.3 and is currently set by Gnus
and used by org-mode to prevent certain unsafe operations. I think the
intent is to use it for blocking more unsafe operations in the future.